# server/internal/platform/httpserver/apigen/components.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package apigen

// Defines values for FieldErrorCode.
const (
	FieldErrorCodeCommonPassword FieldErrorCode = "common_password"
	FieldErrorCodeContainsURL    FieldErrorCode = "contains_url"
	FieldErrorCodeInvalidFormat  FieldErrorCode = "invalid_format"
	FieldErrorCodeMustBeFuture   FieldErrorCode = "must_be_future"
	FieldErrorCodeNotAllowed     FieldErrorCode = "not_allowed"
	FieldErrorCodeOutOfRange     FieldErrorCode = "out_of_range"
	FieldErrorCodeRequired       FieldErrorCode = "required"
	FieldErrorCodeTooLong        FieldErrorCode = "too_long"
	FieldErrorCodeTooShort       FieldErrorCode = "too_short"
	FieldErrorCodeWeakPassword   FieldErrorCode = "weak_password"
)

// Valid indicates whether the value is a known member of the FieldErrorCode enum.
func (e FieldErrorCode) Valid() bool {
	switch e {
	case FieldErrorCodeCommonPassword:
		return true
	case FieldErrorCodeContainsURL:
		return true
	case FieldErrorCodeInvalidFormat:
		return true
	case FieldErrorCodeMustBeFuture:
		return true
	case FieldErrorCodeNotAllowed:
		return true
	case FieldErrorCodeOutOfRange:
		return true
	case FieldErrorCodeRequired:
		return true
	case FieldErrorCodeTooLong:
		return true
	case FieldErrorCodeTooShort:
		return true
	case FieldErrorCodeWeakPassword:
		return true
	default:
		return false
	}
}

// FieldError One invalid field of a request. Clients show text looked up by `code`; `message` is an English explanation for developers. Must match httpserver.FieldError and the field codes of internal/shared.
type FieldError struct {
	// Code What is wrong with the field.
	Code FieldErrorCode `json:"code"`

	// Field The field's path in the request body, e.g. password or tags[1].name.
	Field   string `json:"field"`
	Message string `json:"message"`
}

// FieldErrorCode What is wrong with the field.
type FieldErrorCode string

// Problem RFC 9457 problem details (v0 design 3.5). `title` is the HTTP status phrase, `detail` explains this occurrence, and clients branch on `code`. Must match httpserver.Problem; the platform's contract test checks it.
type Problem struct {
	// Code Stable error code. Platform codes have no prefix (bad_request, unauthorized, not_found, payload_too_large, validation_failed, rate_limited, server_busy, internal_error, not_ready); module codes are prefixed with the module, e.g. identity.email_taken. Each operation lists the codes it can answer in x-problem-codes.
	Code string `json:"code"`

	// Detail What went wrong in this occurrence.
	Detail *string `json:"detail,omitempty"`

	// Errors The invalid fields of the request.
	Errors *[]FieldError `json:"errors,omitempty"`

	// Status HTTP status code.
	Status int `json:"status"`

	// Title HTTP status phrase, e.g. "Not Found".
	Title string `json:"title"`
}